mapping integer to Integer instead of Double in generated documentation
https://github.com/aws-cloudformation/cloudformation-cli/pull/367/files#r465968007

+++ src/rpdk/core/project.py
@@ -75,7 +75,7 @@ SETTINGS_VALIDATOR = Draft7Validator(
 BASIC_TYPE_MAPPINGS = {
     "string": "String",
     "number": "Double",
-    "integer": "Double",
+    "integer": "Integer",
     "boolean": "Boolean",
 }
